+++ Mile.Helpers/Mile.Helpers.Base.cpp
@@ -827,3 +827,22 @@ EXTERN_C BOOL WINAPI MileSetFileAttributesByHandle(
         &BasicInfo,
         sizeof(FILE_BASIC_INFO));
 }
+
+EXTERN_C BOOL WINAPI MileGetFileHardlinkCountByHandle(
+    _In_ HANDLE FileHandle,
+    _Out_ PDWORD HardlinkCount)
+{
+    FILE_STANDARD_INFO StandardInfo;
+
+    BOOL Result = ::GetFileInformationByHandleEx(
+        FileHandle,
+        FILE_INFO_BY_HANDLE_CLASS::FileStandardInfo,
+        &StandardInfo,
+        sizeof(FILE_STANDARD_INFO));
+
+    *HardlinkCount = Result
+        ? StandardInfo.NumberOfLinks
+        : static_cast<DWORD>(-1);
+
+    return Result;
+}